fix(ttyd): delete `MustWait` for removal of resize overlay.
NOTE: This requires the latest `ttyd` version (otherwise the resize
overlay will still be present)

On the latest (master) of `ttyd`, the `disableResizeOverlay` client
options prevents the resize overlay from appearing.

This allows us to remove code which previously waited for the overlay to
show up and then automatically deleted it.