Fix alt modifier
The alt modifier was not handled correctly for both sending and
recording key presses.

To send the key press, the `ESC` sequences needs to be transmitted
before the key.

For recording, the escape sequences lookup table had to be updated with
all alphanumeric combinations.

Syntax highlights had also not been updated.

Signed-off-by: Michael Lorant <michael.lorant@nine.com.au>