Update themes from upstream
Change the update source to the source to the same one that
windowsterminalthemes.dev uses rather than the backup theme file
which is only used if that file cannot be loaded.

https://github.com/atomcorp/themes/blob/master/app/src/App.tsx#L18

Filter out the cursor, selection and credits fields from the source file
to make the diff easier to grok (added back in the following commit).

BREAKING CHANGES:
  - Gruvbox Dark renamed to GruvboxDark.
  - Monokai Octagon renamed to Monokai Pro (Filter Octagon)