Update themes (#377)
* Make themes.json diffs easier to read

This pretty prints the theme file rather than compressing it on to a
single line. This makes it easier to see what has changed between
versions.

* Update themes from upstream

Change the update source to the source to the same one that
windowsterminalthemes.dev uses rather than the backup theme file
which is only used if that file cannot be loaded.

https://github.com/atomcorp/themes/blob/master/app/src/App.tsx#L18

Filter out the cursor, selection and credits fields from the source file
to make the diff easier to grok (added back in the following commit).

BREAKING CHANGES:
  - Gruvbox Dark renamed to GruvboxDark.
  - Monokai Octagon renamed to Monokai Pro (Filter Octagon)

* Add back previously removed theme fields

In the previous commit, the cursor and selection colors were removed, as
well as the credits field to make the commit easier to view. This commit
adds them back.

* Remove themes_custom.json

The themes which were in themes_custom.json are now in the main
themes.json file. This commit removes the themes_custom.json file,
and updates the Makefile and THEMES.md file to reflect this change.