Add initial implementation of keystroke.go
This currently includes a few different things:
- KeyStrokeEvent: Mostly taken from steno almost verbatim.
- KeyStrokeEvents: A wrapper around a slice of KeyStrokeEvents that helps us
package a little state with the events such that we can record exactly when they
should be rendered.
- Page/Keyboard: Another wrapper type around their equivalently named rod types.
These help us ensure that we continue to log the keypresses without substantial
changes to the call sites, and also minimizes the chance of bugs when command
execution is modified since all inputs go through these types.