Tune up the keystroke symbols
Does some tune-up of our special symbols:
- Delete now uses its correct symbol
- PageUp & PageDown is replaced with their correct symbols
- Escape replaced with left diagonal arrow (way more readable too)
- Alt replaced with the option key symbol, not sure how I feel about this.
- Control replaced with ^.
- Shift added.

It should be noted that "correct" here is a bit misleading. I don't know if
there are any 'standard' symbols for some of these keystrokes, but I'm just
choosing ones that seem to map to the key when you search them up online.